Update the OIDC URL (#5)
Update the URL to be used to initiate the OpenID Connect request. Reduce the URL (if no redirect URL is supplied, Azure AD chooses a random one from those entered - only one of which has been), omit the State param as it is unneeded, specify openid instead of the graph scope, and use "organizations" instead of "common" pseudo-tenant, since the configuration for the app registration was limited to only allow org accounts (and common accommodates org and MSA accounts.) Also moved client ID to the end of the string and replaced the sample GUID with an indicator to place their own client ID.

Also includes some small grammar fixes (related to id vs ID).